Fix the pointer-vs-integer size issue the better way.

+++ software/uicomponents/abstract_list_model.cpp
@@ -1,5 +1,6 @@
 #include "uicomponents/abstract_list_model.h"
 #include <cassert>
+#include <stdint.h>
 
 abstract_list_model::abstract_list_model() : stamp(1) {
 }
@@ -51,7 +52,7 @@ GType abstract_list_model::get_column_type_vfunc(int col) const {
 
 bool abstract_list_model::iter_next_vfunc(const iterator &iter, iterator &next) const {
 	assert(iter_is_valid(iter));
-	unsigned long int nextindex = reinterpret_cast<unsigned long int>(iter.gobj()->user_data) + 1U;
+	unsigned int nextindex = static_cast<unsigned int>(reinterpret_cast<uintptr_t>(iter.gobj()->user_data)) + 1U;
 	if (nextindex < alm_rows()) {
 		make_iter_valid(next, nextindex);
 		return true;
@@ -111,12 +112,12 @@ int abstract_list_model::iter_n_root_children_vfunc() const {
 Gtk::TreeModel::Path abstract_list_model::get_path_vfunc(const iterator &iter) const {
 	assert(iter_is_valid(iter));
 	Path p;
-	p.push_back(reinterpret_cast<unsigned long int>(iter.gobj()->user_data));
+	p.push_back(static_cast<unsigned int>(reinterpret_cast<uintptr_t>(iter.gobj()->user_data)));
 	return p;
 }
 
 bool abstract_list_model::iter_is_valid(const iterator &iter) const {
-	return iter.get_stamp() == stamp && reinterpret_cast<unsigned long int>(iter.gobj()->user_data) < alm_rows();
+	return iter.get_stamp() == stamp && static_cast<unsigned int>(reinterpret_cast<uintptr_t>(iter.gobj()->user_data)) < alm_rows();
 }
 
 void abstract_list_model::make_iter_valid(iterator &iter, unsigned int index) const {
@@ -132,14 +133,14 @@ void abstract_list_model::make_iter_invalid(iterator &iter) {
 void abstract_list_model::get_value_vfunc(const iterator &iter, int col, Glib::ValueBase &value) const {
 	assert(iter_is_valid(iter));
 	assert(is_valid_column(col));
-	unsigned long int index = reinterpret_cast<unsigned long int>(iter.gobj()->user_data);
+	unsigned long int index = static_cast<unsigned int>(reinterpret_cast<uintptr_t>(iter.gobj()->user_data));
 	alm_get_value(index, col, value);
 }
 
 void abstract_list_model::set_value_impl(const iterator &iter, int col, const Glib::ValueBase &value) {
 	assert(iter_is_valid(iter));
 	assert(is_valid_column(col));
-	unsigned long int index = reinterpret_cast<unsigned long int>(iter.gobj()->user_data);
+	unsigned long int index = static_cast<unsigned int>(reinterpret_cast<uintptr_t>(iter.gobj()->user_data));
 	alm_set_value(index, col, value);
 }
